refactor sidetree client to be internal module
