+++ Cargo.toml
@@ -26,16 +26,20 @@ web-requests = ["reqwest"]
 async-std = "1.12.0"
 async-trait = "0.1.31"
 base64 = "0.13.0"
+bitflags = "1.2"
 cfg-if = "0.1"
 env_logger = "0.7.1"
+getrandom = { version = "0.2", features = ["js"] }
 log = "0.4.17" # pinned version for dependency of `async-std`
+libsecp256k1 = "0.5"
+multihash = "0.13"
 regex = { version = "1.3.7" }
 reqwest = { version = "0.11", default-features = false, features = ["json", "rustls-tls"], optional = true }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-vade-sidetree-client = { path = "vade-sidetree-client", version = "0.1.0" }
 sha2 = "0.9.5"
 vade = "0.1.0"
+serde_jcs = "0.1"
 
 [dev-dependencies]
 tokio = { version = "=1.7.1", features = ["macros", "rt-multi-thread"] }
