+++ src/vade_sidetree_client/encoder.rs
