+++ src/vade_sidetree_client/mod.rs
@@ -1,3 +1,9 @@
+pub mod did;
+mod encoder;
+pub mod multihash;
+pub mod operations;
+pub mod secp256k1;
+
 use did::{Document, JsonWebKey, JsonWebKeyPublic, PublicKey, Service};
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
@@ -101,8 +107,4 @@ impl fmt::Display for Error<'_> {
         write!(f, "{:?}", self)
     }
 }
-pub mod did;
-mod encoder;
-pub mod multihash;
-pub mod operations;
-pub mod secp256k1;
+
