--- vade-sidetree-client/.gitignore
@@ -1,3 +1,0 @@
-/target
-.DS_Store
-Cargo.lock
