--- vade-sidetree-client/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "vade-sidetree-client"
-version = "0.1.0"
-authors = ["Tomislav Markovski <tomislav@trinsic.id>"]
-edition = "2018"
-description = "Client library for managing DID operations for Sidetree API services"
-license = "Apache-2.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde_jcs = "0.1"
-libsecp256k1 = "0.5"
-getrandom = { version = "0.2", features = ["js"] }
-sha2 = "0.9.5"
-multihash = "0.13"
-base64 = "0.13"
-bitflags = "1.2"
-
-[dependencies.serde]
-default-features = false
-version = "1.0"
-features = ["derive"]
-
-[dependencies.serde_json]
-default-features = false
-features = ["std", "float_roundtrip"]
-version = "1.0"
-
-
-[lib]
-name = "vade_sidetree_client"
-path = "src/lib.rs"
